+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "magic-graph",
+  "name": "magic-graphql",
   "version": "0.0.1",
   "description": "Automatic graph generator for Sequelize ORM, it's magic! 🧙‍♂️🧑🏻‍💻",
   "main": "index.js",
@@ -7,7 +7,9 @@
   "author": "Antonio Orozco <orozco6453@gmail.com>",
   "license": "MIT",
   "private": false,
-  "scripts": {},
+  "scripts": {
+    "build": "rm -rf dist && babel ./src --out-dir dist --ignore 'node_modules'"
+  },
   "dependencies": {
     "apollo-server-express": "^2.17.0",
     "dataloader": "^2.0.0",
@@ -21,5 +23,20 @@
     "sequelize": "^6.3.5",
     "shimmer": "^1.2.1"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@babel/cli": "^7.6.4",
+    "@babel/core": "^7.7.0",
+    "@babel/node": "^7.6.3",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.6.3",
+    "@babel/runtime": "^7.6.3"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env"
+    ],
+    "plugins": [
+      "@babel/plugin-transform-runtime"
+    ]
+  }
 }
